+++ HelloWorldProject/src/main/webapp/WEB-INF/views/home.jsp
@@ -10,5 +10,9 @@
 </h1>
 
 서버로부터 받은 메시지 : ${resultMsg}
+
+<form action="/test">
+	<input type="button" name="postTestBtn" value="POST MAPPING TEST">
+</form>
 </body>
 </html>
